nt_time

Struct FileTime

pub struct FileTime(/* private fields */);

FileTime is a type that represents a Windows file time.

This is a 64-bit unsigned integer value that represents the number of 100-nanosecond intervals that have elapsed since “1601-01-01 00:00:00 UTC”, and is used as timestamps such as NTFS and 7z.

This represents the same value as the FILETIME structure of the Win32 API, which represents a 64-bit unsigned integer value. Note that the maximum value of the FILETIME structure that can be input to the FileTimeToSystemTime function of the Win32 API is limited to “+30828-09-14 02:48:05.477580700 UTC”, which is equivalent to i64::MAX.

Implementations

impl FileTime

pub const NT_TIME_EPOCH: Self

The NT time epoch.

This is defined as “1601-01-01 00:00:00 UTC”.

Examples

assert_eq!(FileTime::NT_TIME_EPOCH, datetime!(1601-01-01 00:00 UTC));

pub const UNIX_EPOCH: Self

The Unix epoch.

This is defined as “1970-01-01 00:00:00 UTC”.

Examples

assert_eq!(FileTime::UNIX_EPOCH, OffsetDateTime::UNIX_EPOCH);

pub const MAX: Self

The largest value that can be represented by the file time.

This is “+60056-05-28 05:36:10.955161500 UTC”.

Examples

assert_eq!(
    FileTime::MAX,
    datetime!(+60056-05-28 05:36:10.955_161_500 UTC)
);

impl FileTime

pub fn to_dos_date_time( self, offset: Option<UtcOffset>, ) -> Result<(u16, u16, u8, Option<UtcOffset>), DosDateTimeRangeError>

Returns MS-DOS date and time which represents the same date and time as this FileTime. This date and time is used as the timestamp such as FAT, exFAT or ZIP file format.

This method returns a (date, time, resolution, offset) tuple.

date and time represents the local date and time. This date and time has no notion of time zone. The resolution of MS-DOS date and time is 2 seconds, but additional finer resolution (10 ms units) can be provided. resolution represents this additional finer resolution.

When the offset parameter is Some, converts date and time from UTC to the local date and time in the provided time zone and returns it with the UTC offset. When the offset parameter is None or is not a multiple of 15 minute intervals, returns the UTC date and time as a date and time and None as the UTC offset.

Errors

Returns Err if the resulting date and time is out of range for MS-DOS date and time.

Panics

Panics if offset is not in the range “UTC-16:00” to “UTC+15:45”.1

Examples

// `1980-01-01 00:00:00 UTC`.
assert_eq!(
    FileTime::new(119_600_064_000_000_000)
        .to_dos_date_time(None)
        .unwrap(),
    (0x0021, u16::MIN, u8::MIN, None)
);
// `2107-12-31 23:59:59 UTC`.
assert_eq!(
    FileTime::new(159_992_927_990_000_000)
        .to_dos_date_time(None)
        .unwrap(),
    (0xff9f, 0xbf7d, 100, None)
);

// Before `1980-01-01 00:00:00 UTC`.
assert!(FileTime::new(119_600_063_990_000_000)
    .to_dos_date_time(None)
    .is_err());
// After `2107-12-31 23:59:59.990000000 UTC`.
assert!(FileTime::new(159_992_928_000_000_000)
    .to_dos_date_time(None)
    .is_err());

// From `2002-11-27 03:25:00 UTC` to `2002-11-26 19:25:00 -08:00`.
assert_eq!(
    FileTime::new(126_828_411_000_000_000)
        .to_dos_date_time(Some(offset!(-08:00)))
        .unwrap(),
    (0x2d7a, 0x9b20, u8::MIN, Some(offset!(-08:00)))
);

When the UTC offset is not a multiple of 15 minute intervals, returns the UTC date and time:

// `2002-11-27 03:25:00 UTC`.
assert_eq!(
    FileTime::new(126_828_411_000_000_000)
        .to_dos_date_time(Some(offset!(-08:01)))
        .unwrap(),
    (0x2d7b, 0x1b20, u8::MIN, None)
);
// `2002-11-27 03:25:00 UTC`.
assert_eq!(
    FileTime::new(126_828_411_000_000_000)
        .to_dos_date_time(Some(offset!(-08:14)))
        .unwrap(),
    (0x2d7b, 0x1b20, u8::MIN, None)
);

// From `2002-11-27 03:25:00 UTC` to `2002-11-26 19:10:00 -08:15`.
assert_eq!(
    FileTime::new(126_828_411_000_000_000)
        .to_dos_date_time(Some(offset!(-08:15)))
        .unwrap(),
    (0x2d7a, 0x9940, u8::MIN, Some(offset!(-08:15)))
);

https://learn.microsoft.com/en-us/windows/win32/fileio/exfat-specification#74101-offsetfromutc-field ↩

pub fn from_dos_date_time( date: u16, time: u16, resolution: Option<u8>, offset: Option<UtcOffset>, ) -> Result<Self, ComponentRange>

Creates a FileTime with the given MS-DOS date and time. This date and time is used as the timestamp such as FAT, exFAT or ZIP file format.

When resolution is Some, additional finer resolution (10 ms units) is added to time.

When offset is Some, converts date and time from the local date and time in the provided time zone to UTC. When offset is None or is not a multiple of 15 minute intervals, assumes the provided date and time is in UTC.

Errors

Returns Err if date or time is an invalid date and time.

Panics

Panics if any of the following are true:

resolution is greater than 199.

offset is not in the range “UTC-16:00” to “UTC+15:45”.1

Examples

// `1980-01-01 00:00:00 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0x0021, u16::MIN, None, None).unwrap(),
    FileTime::new(119_600_064_000_000_000)
);
// `2107-12-31 23:59:59 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0xff9f, 0xbf7d, Some(100), None).unwrap(),
    FileTime::new(159_992_927_990_000_000)
);

// From `2002-11-26 19:25:00 -08:00` to `2002-11-27 03:25:00 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0x2d7a, 0x9b20, None, Some(offset!(-08:00))).unwrap(),
    FileTime::new(126_828_411_000_000_000)
);

// The Day field is 0.
assert!(FileTime::from_dos_date_time(0x0020, u16::MIN, None, None).is_err());
// The DoubleSeconds field is 30.
assert!(FileTime::from_dos_date_time(0x0021, 0x001e, None, None).is_err());

When the UTC offset is not a multiple of 15 minute intervals, assumes the provided date and time is in UTC:

// From `2002-11-26 19:25:00 -08:01` to `2002-11-26 19:25:00 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0x2d7a, 0x9b20, None, Some(offset!(-08:01))).unwrap(),
    FileTime::new(126_828_123_000_000_000)
);
// From `2002-11-26 19:25:00 -08:14` to `2002-11-26 19:25:00 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0x2d7a, 0x9b20, None, Some(offset!(-08:14))).unwrap(),
    FileTime::new(126_828_123_000_000_000)
);

// From `2002-11-26 19:25:00 -08:15` to `2002-11-27 03:40:00 UTC`.
assert_eq!(
    FileTime::from_dos_date_time(0x2d7a, 0x9b20, None, Some(offset!(-08:15))).unwrap(),
    FileTime::new(126_828_420_000_000_000)
);

Additional finer resolution must be in the range 0 to 199:

let _: FileTime = FileTime::from_dos_date_time(0x0021, u16::MIN, Some(200), None).unwrap();

https://learn.microsoft.com/en-us/windows/win32/fileio/exfat-specification#74101-offsetfromutc-field ↩

impl FileTime

pub fn checked_add(self, rhs: Duration) -> Option<Self>

Computes self + rhs, returning None if overflow occurred. The part of rhs less than 100-nanosecond is truncated.

Examples

assert_eq!(
    FileTime::NT_TIME_EPOCH.checked_add(Duration::from_nanos(1)),
    Some(FileTime::NT_TIME_EPOCH)
);
assert_eq!(
    FileTime::NT_TIME_EPOCH.checked_add(Duration::from_nanos(100)),
    Some(FileTime::new(1))
);

assert_eq!(FileTime::MAX.checked_add(Duration::from_nanos(100)), None);

pub fn checked_sub(self, rhs: Duration) -> Option<Self>

Computes self - rhs, returning None if the result would be negative or if overflow occurred. The part of rhs less than 100-nanosecond is truncated.

Examples

assert_eq!(
    FileTime::MAX.checked_sub(Duration::from_nanos(1)),
    Some(FileTime::MAX)
);
assert_eq!(
    FileTime::MAX.checked_sub(Duration::from_nanos(100)),
    Some(FileTime::new(u64::MAX - 1))
);

assert_eq!(
    FileTime::NT_TIME_EPOCH.checked_sub(Duration::from_nanos(100)),
    None
);

pub fn saturating_add(self, rhs: Duration) -> Self

Computes self + rhs, returning FileTime::MAX if overflow occurred. The part of rhs less than 100-nanosecond is truncated.

Examples

assert_eq!(
    FileTime::NT_TIME_EPOCH.saturating_add(Duration::from_nanos(1)),
    FileTime::NT_TIME_EPOCH
);
assert_eq!(
    FileTime::NT_TIME_EPOCH.saturating_add(Duration::from_nanos(100)),
    FileTime::new(1)
);

assert_eq!(
    FileTime::MAX.saturating_add(Duration::from_nanos(100)),
    FileTime::MAX
);

pub fn saturating_sub(self, rhs: Duration) -> Self

Computes self - rhs, returning FileTime::NT_TIME_EPOCH if the result would be negative or if overflow occurred. The part of rhs less than 100-nanosecond is truncated.

Examples

assert_eq!(
    FileTime::MAX.saturating_sub(Duration::from_nanos(1)),
    FileTime::MAX
);
assert_eq!(
    FileTime::MAX.saturating_sub(Duration::from_nanos(100)),
    FileTime::new(u64::MAX - 1)
);

assert_eq!(
    FileTime::NT_TIME_EPOCH.saturating_sub(Duration::from_nanos(100)),
    FileTime::NT_TIME_EPOCH
);

impl FileTime

pub fn to_unix_time(self) -> i64

Returns Unix time which represents the same date and time as this FileTime.

Examples

assert_eq!(FileTime::NT_TIME_EPOCH.to_unix_time(), -11_644_473_600);
assert_eq!(FileTime::UNIX_EPOCH.to_unix_time(), i64::default());
assert_eq!(FileTime::MAX.to_unix_time(), 1_833_029_933_770);

pub fn to_unix_time_nanos(self) -> i128

Returns Unix time in nanoseconds which represents the same date and time as this FileTime.

Examples

assert_eq!(
    FileTime::NT_TIME_EPOCH.to_unix_time_nanos(),
    -11_644_473_600_000_000_000
);
assert_eq!(FileTime::UNIX_EPOCH.to_unix_time_nanos(), i128::default());
assert_eq!(
    FileTime::MAX.to_unix_time_nanos(),
    1_833_029_933_770_955_161_500
);

pub fn from_unix_time(timestamp: i64) -> Result<Self, FileTimeRangeError>

Creates a FileTime with the given Unix time.

Errors

Returns Err if time is out of range for the file time.

Examples

assert_eq!(
    FileTime::from_unix_time(-11_644_473_600).unwrap(),
    FileTime::NT_TIME_EPOCH
);
assert_eq!(
    FileTime::from_unix_time(i64::default()).unwrap(),
    FileTime::UNIX_EPOCH
);
assert_eq!(
    FileTime::from_unix_time(1_833_029_933_770).unwrap(),
    FileTime::MAX - Duration::from_nanos(955_161_500)
);

// Before `1601-01-01 00:00:00 UTC`.
assert!(FileTime::from_unix_time(-11_644_473_601).is_err());
// After `+60056-05-28 05:36:10.955161500 UTC`.
assert!(FileTime::from_unix_time(1_833_029_933_771).is_err());

pub fn from_unix_time_nanos(timestamp: i128) -> Result<Self, FileTimeRangeError>

Creates a FileTime with the given Unix time in nanoseconds.

Errors

Returns Err if time is out of range for the file time.

Examples

assert_eq!(
    FileTime::from_unix_time_nanos(-11_644_473_600_000_000_000).unwrap(),
    FileTime::NT_TIME_EPOCH
);
assert_eq!(
    FileTime::from_unix_time_nanos(i128::default()).unwrap(),
    FileTime::UNIX_EPOCH
);
assert_eq!(
    FileTime::from_unix_time_nanos(1_833_029_933_770_955_161_500).unwrap(),
    FileTime::MAX
);

// Before `1601-01-01 00:00:00 UTC`.
assert!(FileTime::from_unix_time_nanos(-11_644_473_600_000_000_100).is_err());
// After `+60056-05-28 05:36:10.955161500 UTC`.
assert!(FileTime::from_unix_time_nanos(1_833_029_933_770_955_161_501).is_err());

impl FileTime

pub fn now() -> Self

Returns the file time corresponding to “now”.

Panics

Panics if “now” is out of range for the file time.

Examples

let now = FileTime::now();

pub const fn new(ft: u64) -> Self

Creates a new FileTime with the given file time.

Examples

assert_eq!(FileTime::new(u64::MIN), FileTime::NT_TIME_EPOCH);
assert_eq!(FileTime::new(116_444_736_000_000_000), FileTime::UNIX_EPOCH);
assert_eq!(FileTime::new(u64::MAX), FileTime::MAX);

pub const fn to_raw(self) -> u64

Returns the contents of this FileTime as the underlying u64 value.

Examples

assert_eq!(FileTime::NT_TIME_EPOCH.to_raw(), u64::MIN);
assert_eq!(FileTime::UNIX_EPOCH.to_raw(), 116_444_736_000_000_000);
assert_eq!(FileTime::MAX.to_raw(), u64::MAX);

pub const fn to_be_bytes(self) -> [u8; 8]

Returns the memory representation of this FileTime as a byte array in big-endian byte order.

Examples

assert_eq!(FileTime::NT_TIME_EPOCH.to_be_bytes(), [u8::MIN; 8]);
assert_eq!(
    FileTime::UNIX_EPOCH.to_be_bytes(),
    [0x01, 0x9d, 0xb1, 0xde, 0xd5, 0x3e, 0x80, 0x00]
);
assert_eq!(FileTime::MAX.to_be_bytes(), [u8::MAX; 8]);

pub const fn to_le_bytes(self) -> [u8; 8]

Returns the memory representation of this FileTime as a byte array in little-endian byte order.

Examples

assert_eq!(FileTime::NT_TIME_EPOCH.to_le_bytes(), [u8::MIN; 8]);
assert_eq!(
    FileTime::UNIX_EPOCH.to_le_bytes(),
    [0x00, 0x80, 0x3e, 0xd5, 0xde, 0xb1, 0x9d, 0x01]
);
assert_eq!(FileTime::MAX.to_le_bytes(), [u8::MAX; 8]);

pub const fn from_be_bytes(bytes: [u8; 8]) -> Self

Creates a native endian FileTime value from its representation as a byte array in big-endian.

Examples

assert_eq!(
    FileTime::from_be_bytes([u8::MIN; 8]),
    FileTime::NT_TIME_EPOCH
);
assert_eq!(
    FileTime::from_be_bytes([0x01, 0x9d, 0xb1, 0xde, 0xd5, 0x3e, 0x80, 0x00]),
    FileTime::UNIX_EPOCH
);
assert_eq!(FileTime::from_be_bytes([u8::MAX; 8]), FileTime::MAX);

pub const fn from_le_bytes(bytes: [u8; 8]) -> Self

Creates a native endian FileTime value from its representation as a byte array in little-endian.

Examples

assert_eq!(
    FileTime::from_le_bytes([u8::MIN; 8]),
    FileTime::NT_TIME_EPOCH
);
assert_eq!(
    FileTime::from_le_bytes([0x00, 0x80, 0x3e, 0xd5, 0xde, 0xb1, 0x9d, 0x01]),
    FileTime::UNIX_EPOCH
);
assert_eq!(FileTime::from_le_bytes([u8::MAX; 8]), FileTime::MAX);